ExRabbitMQ v2.2.1 ExRabbitMQ.AST.Consumer.GenServer

AST holding module for the consumer behaviour when the surrounding consumer is a GenServer.

Functions

ast()

Produces part of the AST for the consumer behaviour when the consumer is a GenServer.

It holds GenServer handle_info callbacks and a few default implementations.

Specifically, it handles the basic_deliver and basic_cancel AMQP events.

It also responds to connection and channel events, trying to keep a channel open when a connection is available.