ExRabbitMQ v2.2.1 ExRabbitMQ.Connection

A GenServer implementing a long running connection to a RabbitMQ server, with embedded PubSub using :pg2 and :ets.

Functions

close(connection_pid)

close(pid) :: :ok

Gracefully closes the RabbitMQ connection and terminates its GenServer handler identified by connection_pid.

get(connection_pid)

get(pid) :: true | false | {:error, any}

Checks whether this process holds a usable connection to RabbitMQ.

connection_pid is the GenServer pid implementing the called ExRabbitMQ.Connection)

subscribe(connection_pid, connection_config)

subscribe(pid, term) :: true | false

Subscribes a consumer process, via self(), to the managed ETS table.

If the ETS table already contains 65535 consumers, and thus the maximum allowed 65535 channels, then the subscription is not allowed so that a new connection can be created.

connection_pid is the GenServer pid implementing the called ExRabbitMQ.Connection

connection_config is the connection config that the ExRabbitMQ.Connection has to be using in order to allow the subscription