ExRabbitMQ v2.2.0 ExRabbitMQ.Constants

A module holding constants (e.g., process dictionay keys) that are used in multiple sites.

Functions

channel_key()

The key in the process dictionary holding the channel struct used by a consumer or a producer.

channel_monitor_key()

The key in the process dictionary holding the channel monitor reference for the open channel used by a consumer or a producer.

connection_config_key()

The key in the process dictionary holding the connection configuration used by consumers and producers.

connection_pid_key()

The key in the process dictionary holding the connection pid (GenServer) used by a consumer or a producer.

connection_pids_group_name()

The :pg2 group name for the connections to RabbitMQ

no_channel_error()

The error returned when an attempt to use a closed channel is made.

queue_config_key()

The key in the process dictionary holding the queue configuration used by a consumer.