ExRabbitMQ v2.4.0 ExRabbitMQ.Producer behaviour

A behaviour module that abstracts away the handling of RabbitMQ connections and channels.

It also provides hooks to allow the programmer to publish a message without having to directly access the AMPQ interfaces.

For a connection configuration example see ExRabbitMQ.ConnectionConfig.

Example usage for a producer implementing a GenServer

defmodule MyExRabbitMQProducer do
  @module __MODULE__

  use GenServer
  use ExRabbitMQ.Producer

  def start_link() do
    GenServer.start_link(@module, :ok)
  end

  def init(:ok) do
    new_state =
      xrmq_init(:connection, state)
      |> xrmq_extract_state()

    {:ok, new_state}
  end

  def handle_cast({:publish, something}, state) do
    xrmq_basic_publish(something, "", "my_queue")

    {:noreply, state}
  end

  # optional override
  def xrmq_channel_setup(channel, state) do
    # the default channel setup uses the optinal qos_opts from
    # the connection's configuration so we can use it
    # automatically by calling super,
    # unless we want to override everything
    {:ok, new_state} = super(channel, state)

    # any other channel setup goes here...

    {:ok, new_state}
  end
end

Callbacks

xrmq_basic_publish(payload, exchange, routing_key, opts)

xrmq_basic_publish(payload :: term, exchange :: String.t, routing_key :: String.t, opts :: [term]) ::
  :ok |
  {:error, reason :: :blocked | :closing | :no_channel}

This overridable function publishes the payload to the exchange using the provided routing_key.

The wrapper process’s state is passed in to allow the callback to mutate it if overriden.

xrmq_channel_setup(channel, state)

xrmq_channel_setup(channel :: term, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

This hook is called when a connection has been established and a new channel has been opened.

The wrapper process’s state is passed in to allow the callback to mutate it if overriden.

xrmq_extract_state(arg0)

xrmq_extract_state({:ok, state :: term} | {:error, reason :: term, state :: term}) :: state :: term

Helper function that extracts the state argument from the passed in tuple.

xrmq_get_connection_config()

xrmq_get_connection_config() :: term

Returns the connection configuration as it was passed to xrmq_init/2.

This configuration is set in the wrapper process’s dictionary.

For the configuration format see the top section of ExRabbitMQ.Producer.

xrmq_get_env_config(key)

xrmq_get_env_config(key :: atom) :: keyword

Returns a part of the :exrabbitmq configuration section, specified with the key argument.

For the configuration format see the top section of ExRabbitMQ.Producer.

xrmq_init(connection_config, state)

xrmq_init(connection_config :: struct, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

xrmq_init(connection_key :: atom, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

Initiates a connection or reuses an existing one.

When a connection is established then a new channel is opened.

Next, xrmq_channel_setup/2 is called to do any extra work on the opened channel.

This variant accepts a ExRabbitMQ.Connection struct as the argument for the connection_config parameter.

The wrapper process’s state is passed in to allow the callback to mutate it if overriden.

For the configuration format see the top section of ExRabbitMQ.Producer.