ExRabbitMQ v1.1.0 ExRabbitMQ.Consumer behaviour

A behaviour module that abstracts away the handling of RabbitMQ connections and channels.

It abstracts the handling of message delivery and acknowlegement.

It also provides hooks to allow the programmer to wrap the consumption of a message without having to directly access the AMPQ interfaces.

Callbacks

xrmq_basic_ack(delivery_tag, state)

xrmq_basic_ack(delivery_tag :: String.t, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

xrmq_basic_cancel(cancellation_info, state)

xrmq_basic_cancel(cancellation_info :: any, state :: any) ::
  {:noreply, new_state :: term} |
  {:noreply, new_state :: term, timeout | :hibernate} |
  {:noreply, [event :: term], new_state :: term} |
  {:noreply, [event :: term], new_state :: term, :hibernate} |
  {:stop, reason :: term, new_state :: term}

xrmq_basic_deliver(payload, meta, state)

xrmq_basic_deliver(payload :: term, meta :: term, state :: term) ::
  {:noreply, new_state :: term} |
  {:noreply, new_state :: term, timeout | :hibernate} |
  {:noreply, [event :: term], new_state :: term} |
  {:noreply, [event :: term], new_state :: term, :hibernate} |
  {:stop, reason :: term, new_state :: term}

xrmq_basic_reject(delivery_tag, state)

xrmq_basic_reject(delivery_tag :: String.t, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

xrmq_basic_reject(delivery_tag, opts, state)

xrmq_basic_reject(delivery_tag :: String.t, opts :: term, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

xrmq_channel_setup(channel, state)

xrmq_channel_setup(channel :: term, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

xrmq_get_connection_config()

xrmq_get_connection_config() :: term

xrmq_get_env_config(key)

xrmq_get_env_config(key :: atom) :: keyword

xrmq_get_queue_config()

xrmq_get_queue_config() :: term

xrmq_init(connection_config, queue_config, state)

xrmq_init(connection_config :: struct, queue_config :: struct, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

xrmq_init(connection_config :: struct, queue_key :: atom, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

xrmq_init(connection_key :: atom, queue_config :: struct, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

xrmq_init(connection_key :: atom, queue_key :: atom, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}

xrmq_queue_setup(channel, queue, state)

xrmq_queue_setup(channel :: term, queue :: String.t, state :: term) ::
  {:ok, new_state :: term} |
  {:error, reason :: term, new_state :: term}